Rewrited the logic of the settings list in the comment component
