+++ src/feature/post/components/post-list/post/comment/comment-list/comment-item/comment-settings/CommentSettings.tsx
@@ -1,60 +1,34 @@
-import React, { useCallback, useContext } from 'react'
-import { Account, Delete, Update } from '@/components/atoms'
-import { ListItem } from '@/components/list'
+import React, { useContext, useMemo } from 'react'
+import { Div } from '@/components/atoms'
 import SettingsList from '@/components/settings-list'
 import ToggleProvider from '../../../../../../../../../provider/toggle'
 import CommentContext from '../../../../../../../context/comment'
-import CommentListContext from '../../../../../../../context/comment-list'
-import UpdateToggle from '../../../../../../../context/update-toggle'
-import useAppSelector from '../../../../../../../../../hooks/useAppSelector'
-import { authorizationSelector } from '../../../../../../../../../store'
-import { useTranslation } from '../../../../../../../../internationalization'
+import CurrentSettingList from '../../../../../../../context/current-setting-list'
+import CommentSettingsList from './comment-settings-list'
 
 const CommentSettings: React.FC = () => {
-    const user = useAppSelector(authorizationSelector)
-    const { handleToggle } = useContext(UpdateToggle)
-    const { delete: deleteComment } = useContext(CommentListContext)
     const commentContext = useContext(CommentContext)
-    const {translation} = useTranslation()
+    const currentSettingListContext = useContext(CurrentSettingList)
 
-    const handleDeleteComment = useCallback(async () => {
-        if (!commentContext) {
-            return
-        }
-        await deleteComment({
-            commentId: commentContext.data.id.toString()
-        })
-    }, [deleteComment, commentContext?.data.id])
+    const handleSetId = useMemo(() => {
+        return commentContext?.data.id !== undefined 
+            ? currentSettingListContext.handleSetId(commentContext?.data.id) 
+            : undefined
+    }, [commentContext?.data.id, currentSettingListContext.handleSetId])
+
+    const isCurrentComment = currentSettingListContext.id === commentContext?.data.id
 
     return (
         <ToggleProvider>
-            <SettingsList>
-                <ToggleProvider>
-                    <ListItem right={<Account />}>
-                        {translation.feature.post.comment['comment-settings']['list-item'].user}
-                    </ListItem>
-                </ToggleProvider>
-                
-                {
-                    (
-                        user.data?.id === commentContext?.user?.id && 
-                        typeof commentContext?.user?.id === 'number'
-                    ) && (
-                        <>
-                            <ToggleProvider>
-                                <ListItem onClick={handleToggle} right={<Update />}>
-                                    {translation.feature.post.comment['comment-settings']['list-item'].update}
-                                </ListItem>
-                            </ToggleProvider>
-                            <ToggleProvider>
-                                <ListItem onClick={handleDeleteComment} right={<Delete />}>
-                                    {translation.feature.post.comment['comment-settings']['list-item'].delete}
-                                </ListItem>
-                            </ToggleProvider>
-                        </>
-                    )
-                }
-            </SettingsList>
+            <Div onClick={handleSetId}>
+                <SettingsList>
+                    {
+                        isCurrentComment 
+                            ? <CommentSettingsList />
+                            : undefined
+                    }
+                </SettingsList>
+            </Div>
         </ToggleProvider>
     )
 }
