+++ src/feature/post/components/post-list/post/comment/comment-list/comment-item/comment-settings/comment-settings-list/index.ts
@@ -1,0 +1,1 @@
+export {default} from './CommentSettingsList'
